Revert f527f07: insertNewChild/Before at end of line.
Results in the unexpected/counterintuitive need to move the cursor to the beginning of the line before hitting enter if the thought has descendants. This is a common occurrence when editing a thought (with descentants) and then wanting to add a sibling. There is a direct shortcut for insertNewChild/Before, so functional loss, just a learning curve issue. Note that in Google Drive, the behavior is to split the thought from its descendants, even when at the end of the line (so the new thought is empty).