Initial changes for 1098
@raineorshine

FYI the function checks for rootSubthoughts.length > 0, which of course crashes because rootSubthoughts is null, and doesn't have a length field or any field.

I added a test for rootSubthoughts (truthiness = false if null) which fixed this particular crash. But then it crashes on the next tutorial screen, not immediately, but after I follow its instructions, which are to type something.

That crash is taking place in Tutorial2StepContext1.js, where it is also using rootSubthoughts, without considering that it might be null.

We need to decide if the problem is
a) the fact that rootSubthoughts is null, and it shouldn't be
b) it is allowed to be null, but there has to be tests and handling for when it is.

Question: Where is the code that calls Tutorial2StepContext1Parent? This would be the code that could potentially pass it a rootSubthoughts that is not null. I can't find any code that is calling it directly. It gets there through React via the dispatch function, but nowhere else based on the repo files.

One thing that makes this time consuming is that every time I run the project to test it, or run it with some added console.log statements, I have to manually go through all the tutorial steps from the beginning of the first tutorial, after first manually clearing the IndexedDB.