Fix tutorial screens by removing references to `rootSubthoughts`
@raineorshine I fixed a couple of screens by removing references to `rootSubthoughts`, and pushed the results.

Are you ok with this approach? I just want to ensure it's not going to affect other parts of the app I haven't foreseen yet.

Generally, are you comfortable with removing all references to `rootSubthoughts`, including parameters to the Tutorial functions and the `connect(mapStateToProps)` function?

If so, then I can clean up the code and apply it to other screens. However, the inability to get to the next screen causes me to be unable to go through the rest of the tutorial to test and for and fix any additional problems (IE the other ones caused by `rootSubthoughts`). See the screen with the text `'Let's say that you want to make a list of things you have to do at home.`. When adding a thought with the text `To Do` within `Home`, it doesn't advance to the next screen and there is no `Next` button.

This seems to be a distinct bug and should likely be turned into a distinct issue. Let me know if you think we should close this issue and create a new issue for the UI issue in navigating to the next screen.

In the meantime, I'll use the remaining time to tackle the other issue that was assigned.