Remove proxy-memoize
proxy-memoize behaves differently on the dev server vs the build. When memoizing the getSubthoughts selector in the Subthoughts component, it causes caret and rendering issues when deleting thoughts. It is unknown how exactly memoization is different in the build and for what inputs. Maybe Webpack does not have support Proxy or WeakMap. Until this is resolved or the causes are better understood, we need to use a different memoization technique.

Revert "Subthoughts: Memoize children and re-render on change. Fixes hidden thought rendering on load."

This reverts commit d7b277545c8f7e9f0bf90f8bcbfaf5ed5937502c.