Toolbar: Fix toggleHiddenThoughts active state.
Not all ToolbarIcons have the same selector dependencies, so we cannot memoize them the same way. makeBooleanSelector, createSelector, and useCallback add up to a lot of complexity for simple functionality. For now it is removed, but other memoization strategies that hide the complexity and work for all ToolbarIcons will be considered.