Apply autofocus as opacity.
We were unable to use opacity before since it messed up the stacking context. Now that we have a flat list of thoughts instead of a DOM hierarchy this is no longer a problem and we can use opacity on the thought-container and remove specific autofocus styles for the Bullet, Divider, and ThoughtAnnotation.