server: await syncLevelDb in onDocumentLoad to fix synced event.
Prevously, synced was assumed to be broken. This led to a flawed front-end implementation that relied on websocketValueObserved, which may never fire. Without a way to reliably detect when the WebsocketProvider had synced, replicateThought was resolving prematurely on empty idbSynced, causing getDescendantThoughts to incorrectly short circuit. Furthermore, this was only observable on staging where there is server latency.

Properly awaiting initial sync on the server fixes the root issue and allows us to use the synced event on the front-end.