thoughtspace/replicateTree: Do not sync with websocket during export.
Thoughtspace must be replicated before export anyway.

If we were to sync with websocket during export, we should allow replicateThought to resolve as soon as it gets a value from IDB so that the user can still export with a poor connection. But that effectively results in unthrottled calls to bindState on the server as all thoughts are synced.