Re-render VirtualThought only on specific height change.
Previously, all VirtualThoughts were being re-rendered on any height change due to onResize not being a stable object reference. This defeated the purpose of heightRef, which ostensibly allowed VirtualThought height tracking without a re-render. Unfortunately, there seems to be a height discrepancy if a VirtualThought is not re-rendered when its own height changes, though LayoutTree should have everything it needs to re-render from a passive onResize trigger. For now, re-render VirtualThought on its own height change to ensure the thought is rendered with a fresh height.