initEvents/onStateChange: Clear selection when document has lost focus. Closes #1468.
`window.getSelection().focusNode`, `document.activeElement`, and even `visualViewport.height` indicate that the keyboard is open when it is not. Luckily, the invalid state can be detected with `!document.hasFocus()`. This is because (for some reason) the app enters the 'passive' state after hidden -> passive -> active completes.

Since there is no way to force the app to be active when it is passive, then we do not bother trying to re-open the keyboard, and instead disable edit mode so that at least it matches what the user sees.