thoughtspace: Remove forceLocal to ensure that websocket provider gets destroyed.
"Foregrounding" when the thought/lexeme is already loaded only involves updating Redux state, not changing the resolve behavior of background replication. (Note that websocketSynced was already awaited by the time forceLocal was set, so it could not have possibly changed the background replication to the typical foreground replication, which only awaits idbSynced.

forceLocal was mistakenly taken as the opposite of forceRemote.