initialize: Unthrottle updateThoughts until further notice, as it can decrease render performance.
I tried updateThoughtsThrottled with raf-throttle and various _.throttle timeouts in the context of pull. Both resulted in a DECREASE in performance. It is possible that throttling updates may be useful in some contexts, but given the risk of worse performance, it is being removed unless there are measurable improvements and documented constaints.