cursorChanged: Fix accidental conversation of empty string to null (#1769).
The use of || instead of ?? caused the editingValueStore value to change to null instead of empty string on cursor change. This then meant that setCursor incorrectly set state.offsetCursor to null instead of 0 for empty string. This then caused useEditMode to fail when moving the cursor from an empty string.