LayoutTree: Remove extra viewportHeight above while preserving scroll position (#1751).
viewportHeight was previously added to spaceAbove in order to provide enough space to scroll up when spaceAboveDelta is negative. However, that created the reverse problem of adding dead space above when the scroll amount is 0. This commit replaces spaceAbove with spaceAboveExtended which has a minimum height of viewportHeight which can safely be subtracted from the final transformation.