Revert "Debounce scrollCursorIntoView and trigger from LayoutTree effect."
This reverts commit f89c7c5eeba4b57c87b0ff980fba02b974166c05.

scrollCursorIntoView cannot be safely called in an effect, since a re-render can be triggered by a background process such as replication. We should only scroll the user has changed the cursor. Other methods will need to be devised to fix the edge cases that f89c7c5 sought to solve.