Skip throttleReduce tests that are not reliable in CI.
This is still failing intermittently on GitHub Actions even with retries. My guess is that the sub-100ms timing is not reliable in the GitHub Action's virtual compute environment. In practice, the exact number of calls that make it into a given time window is not important (and should not be dependent on). What is important is that all calls are eventually invoked and there is some general concurrency throttling. Given this, and the fact that the tests work perfectly locally, I am going to skip these tests in the CI environment. I will leave the tests here for reference and local testing.