fix: Capacitor build scripts improperly configured
Prior to this fix, `yarn cap:android`  built a static development version of the app. It builds correctly, but upon opening the app the user just gets a black screen.

This was happening because of a line in the Capacitor config file that says to use a development server defined in `.env` instead of loading from local built assets (if `NODE_ENV=development`)

So there was an incongruency – `cap:android` is creating a static build, but the Capacitor config is set up to use live server.

I have resolved this with some changes in `capacitor.config.ts` and some new scripts in `package.json`:

* yarn cap:android:dev – dev build that uses live server
* yarn cap:android:static – dev build that uses static assets
* yarn cap:android – dev build that uses live server (default)

and the same for yarn cap:ios, yarn cap:ios:static, yarn cap:ios:dev.

Now all build scripts work as intended.