taskQueue: Catch rejected taskResolver promises. Fixes #2829.
TODO: catch() is normally placed after then(), but that causes two of the tests to break for some reason.
Furthermore, catch() does not even seem to be called, since adding a console.log does not log anything. How can catch() affect the resolution of a promise that is not rejected?

- restart step completed and total each end
- reset expected and total after end