Enable multiline snapshot tests (baseline).
These snapshots are added to version control as a baseline, but note that the actual layout of multiline thoughts is currently broken.

See: https://github.com/cybersemics/em/issues/2845

The comments such as "TODO: Why are b and c not rendered?" have been removed, as the thoughts in question seem to be rendered fine now. Possibly related to an earlier virtualization bug that has been fixed.