executeCommand: Do not clear selection when restoring the cursor after multiselect.
This behavior is needed for commands like join.

Partial reversion of 07257e7. That commit was intended to fix the cursor after indent/outdent. It's not clear why it needed to clear the selection. It is possible that it was needed at the time, but an interevening change has rendered it unnecessary.