importData: Insert single-line HTML directly at the caret offset.
Copying a single word from a PDF on macOS results in text/html, which by default get processed as multiline.
In order to insert it directly at the caret offset of the cursor thought, we need a special case regex to match single-line content between the body tags. See importData tests for examples.
Otherwise it will be passed to importFiles and import as a child of the current thought.
Eventually importFiles should be modified to insert single-line content at the cursor offset.