fix: only show Cancel command in useFilteredCommands when gesture is in progress
We now use useFilteredCommands to provide command search.

As this hook is also used on mobile to show gesture autocomplete, there was a bug where the cancel command was always being returned and shown in search results on touch devices.

We tighten up the logic in this commit, so that we only show the council command in the gesture autocomplete.