DragAndDropContext: Use PointerTransition instead of MouseTransition (#2967).
MouseTransition fails to properly transition to the TouchBackend on mobile. Either both backends are active, or the TouchBackend is improperly configured. Whaveter the reason, this results in a TouchBackendImpl teardown() every tap, causing the internal handleTopMoveStartDelay to be called twice, overriding the handleTopMoveStart timer, and preventing the first timer from being cleared. That causes beginDrag to fire despite the touchmove.