FadeTransition: Wrap children in a span to allow a default nodeRef.
Optionally override the nodeRef that is passed to CSSTransition. If CSSTransition's nodeRef property is not explicitly provided, it will result in a findDOMNode deprecation warning. In order to avoid making the parent provide the ref every time, we wrap the children in a <span> and use that as the nodeRef by default.

See: https://reactcommunity.org/react-transition-group/transition#Transition-prop-nodeRef