Revert "Continue parent hover highlight pulse after drop (#3215)"
This reverts commit 4ab019689e80e6640fd430e60c118466d01b1f5f.

After real-world use, I've determined that the visual feedback is misleading. The user relies on the pulsing to stop as an indication that the drop completed successfully. We could perhaps pulse in a different color or provide a different signifier to indicate the change, but I will simply revert for now. #2533 helps a lot with this and may be sufficient user feedback anyway.